+++ src/collective/documentgenerator/viewlets/generationlinks.py
@@ -38,10 +38,10 @@ class DocumentGeneratorLinksViewlet(ViewletBase):
             for output_format in template.get_available_formats():
                 title = template.Title()
                 uid = template.UID()
-                link = '{base_url}/document-generation?doc_uid={uid}&output_format={output_format}'.format(
+                link = '{base_url}/document-generation?template_uid={uid}&output_format={output_format}'.format(
                     base_url=base_url,
                     uid=uid,
                     output_format=output_format,
                 )
-                links.append({'link': link, 'title': title, 'output_format': output_format})
+                links.append({'link': link, 'title': title, 'output_format': output_format, 'template_uid': uid})
         return links