+++ docs/CHANGES.rst
@@ -3,7 +3,9 @@ Changelog
 
 0.2 (unreleased)
 ----------------
-
+- Renamed field 'pod_portal_type' to 'pod_portal_types' as this field
+  is a multiselection field.
+  [gbastien]
 - Added field IConfigurablePODTemplate.pod_formats to be able to select the
   format we want to generate the POD template in.
   [gbastien]